requirements.txt
beautifulsoup4==4.12.2
fastapi==0.100.1
python-dotenv==1.0.0
redis==4.6.0
requests==2.31.0
uvicorn==0.23.1
